Improve ProgID retrieval from assoc handlers
As of now, an undocumented interface, IAssocHandlerInfo, is
used to retrieve information about the ProgID. This fixes a
bug present in the previous commit whereby programs with
automatically-generated ProgIDs are not able to be assigned
as a handler for a file type.

This does not mean that any arbitrary association can be
set yet, because we have yet to generate the ProgIDs when
previously-unregistered applications are set, i.e. via the
file browser.

Signed-off-by: Isabella <kawapure@gmail.com>