Correct IAssocHandlerInfo definition
I misread the assembly code when I implemented the last
commit. In reality, the last method did exist, and returned
the flags. Because it had the exact same machine code as
another function, the compiler combined the two, and only
one symbol name was used; I just assumed this was a stub
function for something that was removed.

I spent the better half of 6 hours trying to figure out
what the flags meant here, before I even tried searching it
on GitHub. Karma's a bitch, and it turns out that the enum
definitions for this specific enum were all over GitHub.
That's okay; I came very close to understanding it. If you
are interested, here were the flags as I documented them
before doing the GitHub search:
- 0x10: AUTOMATIC_REGISTRATION
- 0x20: SYSTEM_DEFAULT
- 0x40: REGISTERED_DEFAULT
- 0x80: RECOMMENDED
- 0x200: RECENTLY_INSTALLED

These were the only flags I could find being set in both
shell32.dll and Windows.Storage.dll, so I'm glad I found
the enums on GitHub in any case, because otherwise I would
have spent an inordinate amount of time trying to figure
out where 0x8 comes from, as it is checked for inclusively,
but never set.

I also corrected a small comment mistake in
iassochandlerinfo.h, yay!

Signed-off-by: Isabella <kawapure@gmail.com>