Fix COM embedding startup
It seems that our QueryInterface implementation was the source for most of
the problem. If you didn't understand what went on at a low level, it'd be
practically impossible for you to figure this out, but COM objects work a
little something like this:

- You have a large object composed of a bunch of littler subobjects (think
  of this like object nesting in C). The subobjects always begin with
  virtual function table pointers for that specific subobject, pointing
  back to the parent object. (This happens via thunked IUnknown
  implementations; the compiler generates a new set of IUnknown proxy stubs
  that just call the same function with this shifted, so like
  `return Object::QueryInterface(this - 32, riid, ppvOut);`).
- When QueryInterface is called, you really are supposed to give out the
  pointer of the subobject, not the parent object. The parent object
  is what ultimately implements IUnknown proper, but callers need to access
  the relative offsets in the virtual functions table for their expected
  object.

Unfortunately, C++ as a language obfuscates this a little bit. Casting
objects when multiple inheritence is involved can give you the correct
offset pointer for that subobject, but only if you do the cast right.

Even more unfortunately: although access violations will usually happen,
sometimes calls can succeed and the COM communication will just get stuck
processing the wrong information in the wrong function. This can make
debugging really confusing and difficult.

This is exactly what Aubrey ended up doing in her original QueryInterface
implementation. Because she didn't think of the casting semantics of C++,
she ended up doing an operation equivalent to reinterpret casting the
pointer (the pointer itself not being shifted to point to the interface's
subobject). Even more confusingly, she got extra unlucky and hit one of
those cases where the function call was technically correct and so got hit
with debugging bad input being given to a perfectly fine function.

I replaced the QueryInterface implementation to explicitly cast, which
resolves these issues, at the expense of looking ugly as fuck.

Miscellaneous changes made in this commit:
 - The message loop architecture has been restructured to reside in a
   method of the launcher class. This mirrors the architecture in
   Microsoft's official implementation.
 - I reworked the debugging system for the launcher. It took me a while to
   figure out what exactly was going on, and I cba to remove it now, so the
   ugly debug code muddying up the implementation file is here to stay.
 - I made the launcher implement IExecuteCommand, which was an interface
   Microsoft implemented but Aubrey did not.

Signed-off-by: Isabella <kawapure@gmail.com>