Improve CShellProtectedRegLock reimplementation
I got my hands on private symbols for TWinUI.dll, which allowed me to
change some of the names to ones that align with Microsoft's original
implementation.

I also went ahead and refactored the whole static SID, since that was a
bit of a mess in the original implementation, and it seems that the same
thing can be done in a much simpler manner.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>