Fix default association changing broken for browse items
Previously, it was not possible to change the default association to an
item added via the browse option. This is seemingly because those items
lacked ProgIDs in the first place, and we did not properly handle for this
case.

Since I managed to get ahold of a debug copy of TWinUI with private debug
symbols, this was a piece of cake to reverse engineer and figure out why
Microsoft's implementation was working as intended but our's wasn't.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>