Fix improperly-formed expression fucking up ProgID retrieval
This one was hell to debug.

Of course, I caught that the expression was wrongfully inverted pretty
early on. When I was originally looking at it in IDA, I probably confused
the "!= 0", and that's why I put it in the first place.

The much more nefarious problem is the order-of-operations handling for
bitwise operators in C++. They happen on the same level, from left to
right, so you end up with an expression that always ends up truthy (or in
my case, false, since it was inverted).

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>